Consider logical structure ends as unexecutable
Same reasoning as function declarations. They don't appear in the trace
even when "executed". Only the expressions within then will appear

    if A
    then
      B
    else
      C
    fi

When this expression is executed A, B, and C will appear (B or C
depending on path), but the else, then, and fi will not. The if is
caught by coincidence, but the others we should actively discard as
unexecutable. Similarly for while, for, and case statements.

Fixes #3